+++ CHANGELOG.md
@@ -18,3 +18,6 @@
 # 27/9/2023
 - pick from album and save to local asset db
 - compare local db and rest when open apps after login, when rest add new task, local insert new one. but when rest delete task, local delete too
+# 06/10/2023
+- Uploading File to Server
+- Update local database by ID